Grid system for interactions!
The grid system from my diesel defenders project has been imported to allow for tile interaction by the player a-la stardew valley. There is a stub class for tools that defines a rclick and lclick behavior that tools will inherit in the future.

The indicator box is misaligned with the player and the grid right now, but refining will come with a meeting with the artist.