Quality System
The skeleton of a qualities system is implemented. Ingredients are gathered and have qualities associated with them. These qualities are predefined, but can be combined in ingredients through the gathering process. These qualities will provide flavors to any food item the parent ingredient is used in. Food items are cooked from ingredient items, and draw the ingredient's properties such as qualities and rating.