NPC Road Traversal
NPCs will now spawn from a GuestHandler class and pick a destination across the road. They will spawn with a destination, travel to the destination according to their own copy of the guest script, and will despawn when they pass x = -10. This could lead to some issues, but as the build is stable right now, will refactor this when destinations become more complex.