Pickup and Put Down
Ingredients on the ground will now be picked up and put down on Lclick and Rclick respectively. Items are tracked in a variable on the player, but exist as children of the player node. The ingredient is copied to player and deleted, or copied to the player's position as a sibling and the child copy is deleted.

The reference to the item in code should help with special handling later on, still need to make the placement not on top of the player. Pickup and placement need to correspond to the player's facing, will fix all of these at once along with an animation bug where the player does not turn around when changing direction.

Also, Rclick input is inconsistent, not sure if it's my hardware or a software issue, will return to this.