Item Pickup Refinement: Winner Is You
Finished cleaning up this feature. Added logic to a function outside of the process function for readability and attempted to make objects place in front of player. Need to create a new node to reference interactions with, for now the player will have to be the reference. Created a minigame for this commit to show off the item moving feature.

Winner is you!